Update startup_script.sh
Don't run the installer if CUDA is already installed.

+++ linux/startup_script.sh
@@ -7,5 +7,10 @@ fi
 mkdir -p /opt/google/cuda-installer
 cd /opt/google/cuda-installer/ || exit
 
+if test -f cuda_installation
+then
+  exit
+fi
+
 curl -fSsL -O https://storage.googleapis.com/compute-gpu-installation-us/installer/latest/cuda_installer.pyz
 python3 cuda_installer.pyz install_cuda
